model/requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astroid==2.4.2
astunparse==1.6.3
beautifulsoup4==4.9.3
black==20.8b1
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
isort==5.6.4
jedi==0.17.2
jieba3k==0.35.1
joblib==1.0.0
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
lxml==4.6.2
Markdown==3.3.3
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.8
parso==0.7.1
pathspec==0.8.1
pika==1.1.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
rope==0.18.0
rsa==4.6
sacremoses==0.0.43
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
tinysegmenter==0.3
tldextract==3.1.0
tokenizers==0.9.4
toml==0.10.2
tqdm==4.54.1
transformers==4.1.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wrapt==1.12.1
